MIMA_Project/requirement.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pymupdf>=1.24.0
rank-bm25>=0.2.2
numpy>=1.26.0
sentence-transformers>=3.0.0
transformers>=4.41.0
torch>=2.2.0
scikit-learn>=1.5.0
pillow>=10.0.0
pinecone>=5.0.0
arize-phoenix>=4.0.0
openinference-instrumentation-openai>=0.1.0
ragas>=0.2.0
pandas>=2.2.0